+++ backend/src/main/java/com/yata/backend/domain/yata/repository/yataMemberRepo/YataMemberRepositoryImpl.java
@@ -11,7 +11,6 @@ public class YataMemberRepositoryImpl implements YataMemberRepository {
     private final JPAQueryFactory queryFactory;
     private QYata qYata = QYata.yata;
     private QYataMember qYataMember = QYataMember.yataMember;
-    private QMember qMember = QMember.member;
 
     public YataMemberRepositoryImpl(EntityManager em) {
         this.em = em;
